% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart.R
\name{calculate_heart_intensities}
\alias{calculate_heart_intensities}
\title{calculate_heart_intensities}
\usage{
calculate_heart_intensities(
  heart_raw_data = NULL,
  sleep_data = NULL,
  age = NULL,
  max_hr_formula = "220 - age",
  intensity = NULL,
  intensity_labels = NULL,
  heart_rate_method = 1,
  rest.hr = NULL,
  verbose = FALSE
)
}
\arguments{
\item{heart_raw_data}{Raw heart rate data with time stamp and heart rate columns}

\item{sleep_data}{Sleep data set with startTime and endTime POSIXt time stamps as the columns.}

\item{age}{Age of the participant (extracts from the Fitbit user's profile if NULL), Default: NULL}

\item{max_hr_formula}{An age-based formula to use to calculate maximal heart rate, Default: '220 - age'}

\item{intensity}{Heart rate percentages expressed as decimal points to use when calculating
intensity minutes (uses ACSM as default if NULL), Default: NULL}

\item{intensity_labels}{Labels to go with the `intensity` percentages (uses ACSM default
of sedentary, very light, light, moderate, vigorous, and near maximal if NULL), Default: NULL}

\item{heart_rate_method}{A numeric value of 1, 2, or 3 to specify the approach to calculating heart
rate intensities (see details for more information), Default: 1}

\item{rest.hr}{A constant resting heart rate to use with `heart_rate_method` = 2 if it
is a measure collected by the study, Default: NULL}

\item{verbose}{Print the progress of the Fitbit API heart rate data extraction and calculation of intensity minutes, Default: FALSE}
}
\value{
A data set with sleep, nonwear, and intensity classifications.
}
\description{
Calculates heart rate intensities from raw heart rate data
}
\details{
This function creates heart rate intensities from raw heart rate data.
Sleep time stamps can be passed to remove sleep data from the intensity classification
and customized intensities and intensity labels can be specified if the user chooses
not to use the default ACSM maximal heart rate or heart rate reserve intensities.
}
