% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart.R
\name{extract_heart_intraday}
\alias{extract_heart_intraday}
\title{extract_heart_intraday}
\usage{
extract_heart_intraday(
  token.pathname,
  start.date,
  end.date,
  detail.level,
  imputeZeros = FALSE,
  do.parallel = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{token.pathname}{Full pathname to the location of the access token}

\item{start.date}{The start date specified in the format YYYY-MM-DD or today, Default: Sys.Date()}

\item{end.date}{The end date specified in the format YYYY-MM-DD or today, Default: Sys.Date()}

\item{detail.level}{Detail level of the data points that should be included.
The detail level can be '1sec', '1min', '5min', or '15min', Default: '1min'}

\item{imputeZeros}{Adds zeros to the heart rate variable for missing data}

\item{do.parallel}{Process dates in parallel to speed up extraction of the heart rate data}

\item{verbose}{Print the progress of the Fitbit API heart rate data extraction and calculation of intensity minutes, Default: FALSE}
}
\value{
A data set with time and heart rate variables at the specified interval
}
\description{
Extracts a continuous measure of heart rate data from the Fitbit API
}
\details{
This function retrieves a continuous measure of heart rate from the Fitbit API
at the 1 second, 1 minute, 5 minute, or 15 minute intervals.
}
