% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_fitbit_activities}
\alias{get_fitbit_activities}
\title{Extract Activities from the Fitbit API}
\usage{
get_fitbit_activities(
  token.pathname,
  resource = "All Resources",
  start.date = Sys.Date(),
  end.date = Sys.Date()
)
}
\arguments{
\item{token.pathname}{Path name to the Fitbit API access token.}

\item{resource}{The resource of the data to be returned, Default: "All Resources"}

\item{start.date}{The start date specified in the format YYYY-MM-DD or today, Default: Sys.Date()}

\item{end.date}{The end date specified in the format YYYY-MM-DD or today, Default: Sys.Date()}
}
\value{
Writes the activities to a SQL database stored in the data folder.
}
\description{
Extract calories, distance, elevation, physical activity, and steps from the Fitbit API.
}
\details{
Extract calories, distance, elevation, physical activity, and steps from the Fitbit API.
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[httr]{content}}, \code{\link[httr]{GET}}
 \code{\link[DBI]{dbConnect}}, \code{\link[DBI]{dbWriteTable}}, \code{\link[DBI]{dbDisconnect}}
 \code{\link[RSQLite]{SQLite}}
}
