% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{get_fitbit_device}
\alias{get_fitbit_device}
\title{Extract device information from the Fitbit API}
\usage{
get_fitbit_device(token.pathname, returnData = TRUE, toSQL = FALSE)
}
\arguments{
\item{token.pathname}{Path name to the Fitbit API access token.}

\item{returnData}{Return the data to the user's R environment, Default: TRUE}

\item{toSQL}{Write the data to a SQL database, Default: FALSE}
}
\value{
Writes the device information to a SQL database stored in the data folder.
}
\description{
Extract device, last sync time, and battery level from the Fitbit API.
}
\details{
Extract device, last sync time, and battery level from the Fitbit API.
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[httr]{content}}, \code{\link[httr]{GET}}
 \code{\link[DBI]{dbConnect}}, \code{\link[DBI]{dbWriteTable}}, \code{\link[DBI]{dbDisconnect}}
 \code{\link[RSQLite]{SQLite}}
}
