% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{get_fitbit_exercise_log}
\alias{get_fitbit_exercise_log}
\title{Extract Exercise Log from the Fitbit API}
\usage{
get_fitbit_exercise_log(
  token.pathname,
  after.date = Sys.Date(),
  limit = 100,
  returnData = TRUE,
  toSQL = FALSE
)
}
\arguments{
\item{token.pathname}{Path name to the Fitbit API access token.}

\item{after.date}{The date in the format Y-m-d to limit the extraction of the exercise records.}

\item{limit}{Number of exercise records to extract from the Fitbit API (Maximum: 100), Default: 100}

\item{returnData}{Return the data to the user's R environment, Default: TRUE}

\item{toSQL}{Write the data to a SQL database, Default: FALSE}
}
\value{
Writes the exercise log to a SQL database stored in the data folder.
}
\description{
Extract date, time, type, duration, steps, calories, and heart
    rate in exercise sessions from the Fitbit API.
}
\details{
Extract date, time, type, duration, steps, calories, and heart
    rate in exercise sessions from the Fitbit API.
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}},
 \code{\link[httr]{content}}, \code{\link[httr]{GET}},
 \code{\link[DBI]{dbConnect}}, \code{\link[DBI]{dbWriteTable}}, \code{\link[DBI]{dbDisconnect}},
 \code{\link[RSQLite]{SQLite}}
}
