% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart.R
\name{get_fitbit_heart_intraday}
\alias{get_fitbit_heart_intraday}
\title{FUNCTION_TITLE}
\usage{
get_fitbit_heart_intraday(
  token.pathname,
  start.date = Sys.Date(),
  end.date = Sys.Date(),
  detail.level = "1min"
)
}
\arguments{
\item{token.pathname}{PARAM_DESCRIPTION}

\item{start.date}{PARAM_DESCRIPTION, Default: Sys.Date()}

\item{end.date}{PARAM_DESCRIPTION, Default: Sys.Date()}

\item{detail.level}{PARAM_DESCRIPTION, Default: '1min'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[DBI]{dbConnect}}, \code{\link[DBI]{dbExistsTable}}, \code{\link[DBI]{dbReadTable}}, \code{\link[DBI]{dbWriteTable}}, \code{\link[DBI]{dbDisconnect}}
 \code{\link[RSQLite]{SQLite}}
 \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{case_when}}, \code{\link[dplyr]{summarise_all}}, \code{\link[dplyr]{select}}
 \code{\link[utils]{head}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[httr]{content}}, \code{\link[httr]{GET}}
}
