% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{get_fitbit_profile}
\alias{get_fitbit_profile}
\title{Extract Profile from the Fitbit API}
\usage{
get_fitbit_profile(token.pathname)
}
\arguments{
\item{token.pathname}{Path name to the Fitbit API access token.}
}
\value{
Returns the Fitbit Profile of the user
}
\description{
Extract profile information from the Fitbit API
}
\details{
Extract profile information from the Fitbit API
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[httr]{content}}, \code{\link[httr]{GET}}
}
