% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{get_fitbit_report}
\alias{get_fitbit_report}
\title{get_fitbit_report}
\usage{
get_fitbit_report(
  database_pathname,
  reports_pathname = NULL,
  returnData = TRUE,
  toHTML = FALSE,
  toCSV = FALSE,
  study_name = "study",
  report_author = "iFitbit",
  reportName = "FitbitReport",
  start.date,
  end.date,
  ...
)
}
\arguments{
\item{database_pathname}{Full pathname to the SQL database}

\item{reports_pathname}{Full pathname to the location that the HTML or CSV file should be stored, Default: NULL}

\item{returnData}{Returns the combined data set, Default: TRUE}

\item{toHTML}{Renders a Rmarkdown report to HTML, Default: FALSE}

\item{toCSV}{Writes the full data set to a CSV file, Default: FALSE}

\item{study_name}{A customized study name for the iFitbit report, Default: 'study'}

\item{report_author}{A customized report author for the iFitbit report, Default: 'iFitbit'}

\item{reportName}{PARAM_DESCRIPTION, Default: 'FitbitReport'}

\item{start.date}{Add a start date to allow filtering of data when returning the data or writing it to a CSV}

\item{end.date}{Add an end date to allow filtering of data when returning the data or writing it to a CSV}

\item{...}{Additional arguments passed to \code{\link[data.table]{fwrite}} or \code{\link[rmarkdown]{render}}}
}
\value{
Returns data set or generates a HTML or CSV file
}
\description{
Generates an HTML report, returns data, or writes to a CSV file
}
\details{
Generates an HTML report, returns data, or writes to a CSV file
}
\seealso{
\code{\link[DBI]{dbConnect}}, \code{\link[DBI]{dbReadTable}}, \code{\link[DBI]{dbDisconnect}}
 \code{\link[RSQLite]{SQLite}}
 \code{\link[rmarkdown]{render}}
 \code{\link[data.table]{fwrite}}
}
