% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{get_fitbit_sleep}
\alias{get_fitbit_sleep}
\title{Extract Sleep from the Fitbit API}
\usage{
get_fitbit_sleep(
  token.pathname,
  start.date = Sys.Date(),
  end.date = Sys.Date(),
  toSQL = TRUE,
  returnData = FALSE
)
}
\arguments{
\item{token.pathname}{Path name to the Fitbit API access token.}

\item{start.date}{The start date specified in the format YYYY-MM-DD or today, Default: Sys.Date()}

\item{end.date}{The end date specified in the format YYYY-MM-DD or today, Default: Sys.Date()}

\item{toSQL}{Writes the activities to a SQL database stored in the data folder, Default: TRUE}

\item{returnData}{Returns the sleep data, Default: FALSE}
}
\value{
Output is either written to a SQL database or returned as a dataframe
}
\description{
Extract sleep outcomes including duration, efficiency, time
    asleep, time awake, minutes to fall asleep, and time in bed.
}
\details{
Extract sleep outcomes including duration, efficiency, time
    asleep, time awake, minutes to fall asleep, and time in bed.
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[httr]{content}}, \code{\link[httr]{GET}}
 \code{\link[DBI]{dbConnect}}, \code{\link[DBI]{dbWriteTable}},
 \code{\link[DBI]{dbExistsTable}}, \code{\link[DBI]{dbRemoveTable}},
 \code{\link[DBI]{dbDisconnect}}, \code{\link[RSQLite]{SQLite}}
}
