% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iFitbit.R
\docType{package}
\name{iFitbit}
\alias{iFitbit}
\title{iFitbit: R Package for Extracting Data from the Fitbit API and Generating Interactive Reports}
\description{
This R Package extracts data from the Fitbit API and generates
    interactive reports, tables, and visualizations.
}
\section{iFitbit functions}{


\code{\link{authorize_fitbit_app}}

\code{\link{refresh_fitbit_token}}

\code{\link{get_fitbit_activities}}

\code{\link{get_fitbit_exercise_log}}

\code{\link{get_fitbit_sleep}}

\code{\link{get_fitbit_profile}}

\code{\link{get_fitbit_device}}

\code{\link{get_fitbit_heart_intraday}}
}

