% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{refresh_fitbit_token}
\alias{refresh_fitbit_token}
\title{Refresh Fitbit API Token}
\usage{
refresh_fitbit_token(token.pathname, key, secret = NULL, force.refresh = FALSE)
}
\arguments{
\item{token.pathname}{Path name to the token that needs to be refreshed.}

\item{key}{consumer key, also sometimes called the client ID}

\item{secret}{consumer secret, also sometimes called the client secret.
Despite its name, this does not necessarily need to be protected like a
password, i.e. the user still has to authenticate themselves and grant the
app permission to access resources on their behalf. For example, see
Google's docs for \href{https://developers.google.com/identity/protocols/oauth2}{OAuth2 for installed applications}.}

\item{force.refresh}{Force the refresh token when it has not expired, Default: FALSE}
}
\value{
Refreshes the token retrieved from Fitbit API authorization and saves it to directory.
}
\description{
Refresh the Fitbit API Token after it expires to access Fitbit API
}
\details{
Refresh the Fitbit API Token after it expires to access Fitbit API
}
\seealso{
\code{\link[httr]{add_headers}}
 \code{\link[httr]{content_type}}
 \code{\link[httr]{POST}}
 \code{\link[httr]{content}}
 \code{\link[RCurl]{base64}}
 \code{\link[utils]{modifyList}}
}
