% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_fitbit_exercise_log}
\alias{get_fitbit_exercise_log}
\title{Extract Exercise Log from the Fitbit API}
\usage{
get_fitbit_exercise_log(token.pathname, limit = 25)
}
\arguments{
\item{token.pathname}{Path name to the Fitbit API access token.}

\item{limit}{Number of exercise records to extract from the Fitbit API, Default: 25}
}
\value{
Writes the exercise log to a SQL database stored in the data folder.
}
\description{
Extract date, time, type, duration, steps, calories, and heart
    rate in exercise sessions from the Fitbit API.
}
\details{
Extract date, time, type, duration, steps, calories, and heart
    rate in exercise sessions from the Fitbit API.
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}},
 \code{\link[httr]{content}}, \code{\link[httr]{GET}},
 \code{\link[DBI]{dbConnect}}, \code{\link[DBI]{dbWriteTable}}, \code{\link[DBI]{dbDisconnect}},
 \code{\link[RSQLite]{SQLite}}
}
