% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart.R
\name{get_fitbit_heart_intraday}
\alias{get_fitbit_heart_intraday}
\title{get_fitbit_heart_intraday}
\usage{
get_fitbit_heart_intraday(
  token.pathname,
  start.date = Sys.Date(),
  end.date = Sys.Date(),
  detail.level = "1min",
  returnData = TRUE,
  toSQL = FALSE
)
}
\arguments{
\item{token.pathname}{Full pathname to the location of the access token}

\item{start.date}{The start date specified in the format YYYY-MM-DD or today, Default: Sys.Date()}

\item{end.date}{The end date specified in the format YYYY-MM-DD or today, Default: Sys.Date()}

\item{detail.level}{Detail level of the data points that should be included. The detail level can be '1sec', '1min', '5min', or '15min', Default: '1min'}

\item{returnData}{Return the data to the user's R environment, Default: TRUE}

\item{toSQL}{Write the data to a SQL database, Default: FALSE}
}
\value{
A measure of day-level non-wear, sleep, sedentary time, and physical activity
}
\description{
Retreives a continuous measure of heart rate from the Fitbit API
}
\details{
Retreives a continuous measure of heart rate from the Fitbit API
}
\seealso{
\code{\link[DBI]{dbConnect}}, \code{\link[DBI]{dbExistsTable}}, \code{\link[DBI]{dbReadTable}}, \code{\link[DBI]{dbWriteTable}}, \code{\link[DBI]{dbDisconnect}}
 \code{\link[RSQLite]{SQLite}}
 \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{case_when}}, \code{\link[dplyr]{summarise_all}}, \code{\link[dplyr]{select}}
 \code{\link[utils]{head}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[httr]{content}}, \code{\link[httr]{GET}}
}
